added AuthForm component
