+++ src/components/auth-form/auth-form.module.css
@@ -1,0 +1,15 @@
+.inputs {
+  display: grid;
+  gap: 24px;
+  margin: 0 auto;
+}
+
+.form {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.title {
+  margin: 0;
+}
