+++ src/pages/register/register.module.css
@@ -9,19 +9,3 @@
   display: flex;
   text-decoration: none;
 }
-
-.inputs {
-  display: grid;
-  gap: 24px;
-  margin: 0 auto;
-}
-
-.form {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.title {
-  margin: 0;
-}
